ginkgo: allow statements to sometimes return values really
We want to emulate Rust's sytle of unterminated statements acting more
like expressions. We do this by differentiating expressions that are terminated
from those that aren't, and then checking unterminated ones only appear
in positions they are allowed to. In the future, we might find that this
is better to enforce in the parser rather than the interpreter, now I'm thinking
about it.