ginkgo: start doing basic binding resolution
This highlights a big problem with supporting first-class functions so early:
we've already highlighted a situation where a borrow-checking type of thing
is needed, but this is still a good improvement in correctness for 'normal'
code.

This also adds a wrapping-struct around `Expr` and `Stmt` (now called `ExprTyp` and `StmtTyp`).
This was because I was going to add IDs to them and store the resolution info in
a side table - I've kept it because I still might do that, and because we probs
want to add spans to the AST at some point when we do error reporting. We
may also get rid of that redirection later, unknown.